+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Improve handling of short search terms in project search
 * Add copy button for AI prompts and responses
 * Markdown editor: improve code block highlighting in dark mode
+* Fix marmaid incorrect HTML label rendering in PDF
 
 
 ## v2026.36 - 2026-05-20